apache-proxy/requirements.txt
# Lists the Python dependencies required to run the Apache-backed RedirectIQ benchmark target.
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
bcrypt==4.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
qrcode[pil]==8.0
requests==2.32.3
